pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "noise-suppression"
version = "0.1.0"
description = "Local MVP toolkit for Russian speech enhancement experiments"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "jiwer>=3.0.5",
  "numpy>=2.2.4",
  "pyyaml>=6.0.2",
  "rich>=14.0.0",
  "scipy>=1.15.2",
  "soundfile>=0.13.1",
  "typer>=0.16.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "ruff>=0.11.12",
]
metrics = [
  "pesq>=0.0.4",
  "pystoi>=0.4.1",
]
data = [
  "datasets>=4.0.0",
]
train = [
  "torch>=2.6.0",
]

[project.scripts]
noise-suppression = "noise_suppression.cli:app"

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["notebooks"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
